% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{balanced.folds}
\alias{balanced.folds}
\title{balanced.folds}
\usage{
balanced.folds(y, nfolds = 3)
}
\arguments{
\item{y}{A response vector. If a factor, classification is assumed, otherwise regression is assumed.}

\item{nfolds}{The number of folds in the cross validation.}
}
\value{
folds
}
\description{
Get balanced folds where each fold has close to overall class ratio
}
\examples{
y<-factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15)))
print(balanced.folds(y, nfolds=3))
y<- 1:60
print(balanced.folds(y, nfolds=3))
}
\author{
Shi Huang
}
