% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_reg_plot.R
\name{boxplot_rel_predicted_train_vs_test}
\alias{boxplot_rel_predicted_train_vs_test}
\title{boxplot_rel_predicted_train_vs_test}
\usage{
boxplot_rel_predicted_train_vs_test(relTrain_data,
  train_target_field = "value", train_prefix = "train",
  test_prefix = "test", outdir = NULL)
}
\arguments{
\item{relTrain_data}{The output dataframe of \code{calc_rel_predicted}.}

\item{train_target_field}{A string indicating the target field of the training metadata for regression.}

\item{train_prefix}{The prefix for the dataset in the training data.}

\item{test_prefix}{The prefix for the dataset in the testing data.}

\item{outdir}{The output directory.}
}
\description{
Make the boxplot the relative predicted values in both train and test datasets.
}
\examples{
set.seed(123)
train_x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
train_y<- 1:60
test_x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209)))))
test_y<- 1:45
train_rf_model<-rf.out.of.bag(train_x, train_y)
predicted_test_y<-predict(train_rf_model$rf.model, test_x)$predictions
relTrain_data<-calc_rel_predicted(train_y, train_rf_model$predicted, test_y, predicted_test_y)
boxplot_rel_predicted_train_vs_test(relTrain_data)
}
\author{
Shi Huang
}
