% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trimming_util.R
\name{check_metadata}
\alias{check_metadata}
\title{check_metadata}
\usage{
check_metadata(metadata, more_missing_values = NULL, unique_rate_thres = 0.2)
}
\arguments{
\item{metadata}{A dataframe with > two columns corresponds to samples (rownames) in the biological data.}

\item{more_missing_values}{A optional string(s) can be added to define the missing values.}

\item{unique_rate_thres}{The minimum ratio of unique values to non-missing values for an integer variable to be treated as numeric rather than categorical.}
}
\value{
A data.frame summarizing each of metadata variables.
}
\description{
Summarize the completeness and the data type (numeric or categorical) of each metadata variable.
}
\examples{
set.seed(123)
a<-factor(c(rep("A", 29), NA, rep("B", 29), NA))
b<-factor(c(rep("A", 27), NA, "Not applicable", "Missing:not collected", rep("B", 28), NA, NA))
c<-factor(c(rep("A", 20), rep("B", 20), rep("C", 20)))
d<-c(sample(1:18), NA, "Not applicable", sample(5:44))
e<-c(rnorm(40), NA, "Not applicable", rnorm(18, 4))
e0<-c(NA, "Not applicable", rnorm(58, 4))
f<-rep("C", 60)
g<-rep(4, 60)
h<-c(rep("C", 59), "B")
metadata<-data.frame(a, b, c, d, e, e0, f, g, h)
check_metadata(metadata)
}
\author{
Shi Huang
}
