% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetweenGroup.test.R
\name{desc_stats_all}
\alias{desc_stats_all}
\title{desc_stats_all}
\usage{
desc_stats_all(x, y, positive_class = NA, clr_transform = FALSE)
}
\arguments{
\item{x}{A data.martrix or data.frame including multiple numeric vectors.}

\item{y}{A factor with two or more levels.}

\item{clr_transform}{A logical value indicates if clr transformation before statistical analysis of compositional microbiome data.}
}
\description{
The descriptive statistics summary of all features of compositional microbiome data.
}
\examples{
y <-factor(c(rep("A", 30), rep("B", 30)))
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
desc_stats_all(x, y, clr_transform=FALSE)
desc_stats_all(x, y, clr_transform=TRUE)
}
\author{
Shi Huang
}
