% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trimming_util.R
\name{filter_samples_by_groups_in_target_field_of_metadata}
\alias{filter_samples_by_groups_in_target_field_of_metadata}
\title{filter_samples_by_groups_in_target_field_of_metadata}
\usage{
filter_samples_by_groups_in_target_field_of_metadata(
  data,
  metadata,
  target_field,
  groups,
  negate = FALSE,
  ids_col = NA
)
}
\arguments{
\item{data}{A data.frame.}

\item{metadata}{A data.frame including multiple metadata variables corresponds to samples in the data.}

\item{target_field}{A character string indicating a target variable in the metadata.}

\item{groups}{A character string(s) indicating one or multiple groups in the target metadata variable specified.}

\item{negate}{A bool value indicates if samples in the specified groups should be removed (TRUE) or kept (FALSE).}

\item{ids_col}{The column of metadata holding sample IDs. If NA, the rownames of metadata are used.}
}
\value{
A list with the filtered \code{data} and \code{metadata}.
}
\description{
Keep (or remove) the samples belonging to the specified groups of a metadata variable.
}
\examples{
set.seed(123)
data <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
z<-factor(c(rep("E", 20), rep("C", 20), rep("D", 20)))
y<-factor(c(rep("A", 30), rep("B", 30)))
y0<-factor(c(rep("A", 5), rep("B", 55)))
metadata <- data.frame(z, y, y0)
filter_samples_by_groups_in_target_field_of_metadata(data, metadata, target_field="z",
                                                     groups=c("D", "A"), negate=TRUE)
}
