% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_util.R
\name{generate.comps_datalist}
\alias{generate.comps_datalist}
\title{generate.comps_datalist}
\usage{
generate.comps_datalist(df, f, comp_group)
}
\arguments{
\item{df}{Training data: a data.frame.}

\item{f}{A factor in the metadata with at least two levels (groups).}

\item{comp_group}{A string indicates the group in the f}
}
\value{
A object of list with elements: a list of data.frame and a list of factor
}
\description{
To generate sub-datasets and sub-metadata by pairing
one level and each of other levels of the specified category in the metadata.
}
\author{
Shi Huang
}
