% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{get.auroc}
\alias{get.auroc}
\title{get.auroc}
\usage{
get.auroc(predictor, y, positive_class)
}
\arguments{
\item{predictor}{A predictor. For example, one column in the probability table, suggesting
the probability assign a given observation to the positive class in the factor y.}

\item{y}{A binary factor vector indicates observed values.}

\item{positive_class}{A class of the factor y.}
}
\value{
The auroc value.
}
\description{
calculates the area under the ROC curve
}
\examples{
y<-factor(c(rep("A", 31), rep("B", 29)))
pred <- c(runif(30, 0.5, 0.9), runif(30, 0, 0.6))
prob <-data.frame(A=pred, B=1-pred)
positive_class="A"
get.auroc(predictor=prob[, positive_class], y, positive_class="A")
get.auroc(predictor=prob[, positive_class], y, positive_class="B")
}
\author{
Shi Huang
}
