% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{get.oob.probability.from.forest}
\alias{get.oob.probability.from.forest}
\title{get.oob.probability.from.forest}
\usage{
get.oob.probability.from.forest(model, x)
}
\arguments{
\item{model}{A \code{ranger} classification model trained with \code{keep.inbag=TRUE}.}

\item{x}{The training data.}
}
\value{
A table of probabilities \code{probs}
}
\description{
Get probability of each class using only out-of-bag predictions from RF
}
\author{
Shi Huang
}
