% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{get.reg.oob.performance}
\alias{get.reg.oob.performance}
\title{get.reg.oob.performance}
\usage{
get.reg.oob.performance(model, y)
}
\arguments{
\item{model}{A object of class \code{rf.out.of.bag} with elements.}

\item{y}{The dependent variable in training data.}
}
\value{
\code{perf}
}
\description{
Get oob performance from forests for training data. Last update: Apr. 7, 2019
}
\author{
Shi Huang
}
