% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{get.reg.predict.performance}
\alias{get.reg.predict.performance}
\title{get.reg.predict.performance}
\usage{
get.reg.predict.performance(model, newx, newy)
}
\arguments{
\item{model}{A object of random forest out-of-bag regression model.}

\item{newx}{The external test data.}

\item{newy}{The dependent variable in the external test data.}
}
\value{
\code{perf}
}
\description{
Get prediction performance from forests for the external test data. Last update: Apr. 7, 2019
}
\author{
Shi Huang
}
