% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_plot_util.R
\name{plot.reg_res_list}
\alias{plot.reg_res_list}
\title{plot.reg_res_list}
\usage{
\method{plot}{reg_res_list}(res_list, outdir = "./")
}
\arguments{
\item{outdir}{The output directory.}

\item{reg_res_list}{a list, the output of the function \code{rf_clf.by_datasets}.}
}
\description{
Plot a summary of the performance of a list of classification models
output from \code{rf_clf.by_datasets}.
}
\examples{
df <- data.frame(rbind(t(rmultinom(7, 75, c(.21,.6,.12,.38,.099))),
            t(rmultinom(8, 75, c(.001,.6,.42,.58,.299))),
            t(rmultinom(15, 75, c(.011,.6,.22,.28,.289))),
            t(rmultinom(15, 75, c(.091,.6,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.6,.42,.58,.299)))))
df0 <- data.frame(t(rmultinom(60, 300,c(.001,.6,.2,.3,.299))))
metadata<-data.frame(f_s=factor(c(rep("A", 15), rep("B", 15), rep("A", 15), rep("B", 15))),
                     f_c=factor(c(rep("C", 30), rep("D", 30))),
                     f_d=factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15))),
                     age=c(1:30, 2:31)
                     )
reg_res_list<-rf_reg.by_datasets(df, metadata, s_category="f_c", c_category="age")
plot.reg_res_list(reg_res_list)
}
\seealso{
ranger rf_reg.by_datasets
}
\author{
Shi Huang
}
