% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_plot_util.R
\name{plot_clf_res_list}
\alias{plot_clf_res_list}
\title{plot_clf_res_list}
\usage{
plot_clf_res_list(
  clf_res_list,
  p_cutoff = 0.05,
  p.adj.method = "bonferroni",
  q_cutoff = 0.05,
  outdir = NULL,
  plot_height = NULL
)
}
\arguments{
\item{clf_res_list}{A list, the output of the function \code{rf_clf.by_datasets}.}

\item{p_cutoff}{A number indicating the cutoff of p values.}

\item{p.adj.method}{A string indicating the p-value correction method.}

\item{q_cutoff}{A number indicating the cutoff of q values after fdr correction.}

\item{outdir}{The output directory. If NULL, no file is written.}

\item{plot_height}{The plot height.}
}
\description{
Plot a summary of the performance of a list of classification models
output from \code{rf_clf.by_datasets}.
}
\examples{
df <- data.frame(rbind(t(rmultinom(7, 75, c(.21,.6,.12,.38,.099))),
            t(rmultinom(8, 75, c(.001,.6,.42,.58,.299))),
            t(rmultinom(15, 75, c(.011,.6,.22,.28,.289))),
            t(rmultinom(15, 75, c(.091,.6,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.6,.42,.58,.299)))))
metadata<-data.frame(f_s=factor(c(rep("A", 15), rep("B", 15), rep("A", 15), rep("B", 15))),
                     f_c=factor(c(rep("C", 30), rep("D", 30))),
                     f_d=factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15))),
                     age=c(1:30, 2:31)
                     )
clf_res_list<-rf_clf.by_datasets(df, metadata, s_category="f_c", c_category="f_s")
plot_clf_res_list(clf_res_list)
}
\seealso{
ranger rf_clf.by_datasets
}
\author{
Shi Huang
}
