% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_plot_util.R
\name{plot_logfc_heatmap}
\alias{plot_logfc_heatmap}
\title{plot_logfc_heatmap}
\usage{
plot_logfc_heatmap(
  feature_res,
  logfc_col = "mean_logfc",
  features = NULL,
  min_sig_datasets = 1,
  feature_md = NULL,
  feature_id_col = "Feature_ID",
  label_col = NULL,
  outdir = NULL,
  plot_width = 6,
  plot_height = NULL
)
}
\arguments{
\item{feature_res}{A data.frame of feature statistics stacked over datasets with columns "feature", "dataset",
"Enr" and a log2 fold change column (e.g., the \code{feature_res} of \code{plot_clf_res_list},
\code{rf_clf.by_datasets.summ} or \code{rf_clf.comps.summ}), or an \code{rf_clf.by_datasets}/\code{rf_clf.comps} object.}

\item{logfc_col}{The column of log2 fold changes: "mean_logfc", "median_logfc" or "generalized_logfc".}

\item{features}{An optional vector of features to show. By default, the features significant in
at least \code{min_sig_datasets} datasets.}

\item{min_sig_datasets}{The minimum number of datasets in which a feature is significant to be shown.}

\item{feature_md}{An optional data.frame of feature annotations (e.g., taxonomy) used to label the features.}

\item{feature_id_col}{The feature ID column in \code{feature_md}.}

\item{label_col}{The column in \code{feature_md} used as feature labels.}

\item{outdir}{The output directory. If NULL, no file is written.}

\item{plot_width}{The width (inches) of heatmap.}

\item{plot_height}{The height (inches) of heatmap. By default, it scales with the number of features.}
}
\value{
A list including the matrix of log2 fold changes (features by datasets) and the heatmap.
}
\description{
Heatmap of the log2 fold changes of features across datasets (e.g., cohorts, body sites or timepoints),
with significantly enriched or depleted features marked by asterisks. It facilitates the identification of
generalizable versus study-specific microbial signatures.
}
\examples{
df <- data.frame(rbind(t(rmultinom(30, 300, c(.21,.6,.12,.38,.099))),
                       t(rmultinom(30, 300, c(.001,.6,.42,.58,.299))),
                       t(rmultinom(30, 300, c(.21,.6,.12,.38,.099))),
                       t(rmultinom(30, 300, c(.001,.6,.42,.58,.299)))))
metadata <- data.frame(study=factor(rep(c("S1", "S2"), each=60)),
                       disease=factor(rep(rep(c("Case", "Control"), each=30), 2)))
res <- rf_clf.by_datasets(df, metadata, s_category="study", c_category="disease",
                          positive_class="Case", n_cores=1)
plot_logfc_heatmap(res)$plot
}
