% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_reg_plot.R
\name{plot_obs_VS_pred}
\alias{plot_obs_VS_pred}
\title{plot_obs_VS_pred}
\usage{
plot_obs_VS_pred(y, predicted_y, prefix = "train",
  target_field = "value", span = 1, outdir = NULL)
}
\arguments{
\item{y}{The numeric values for labeling data.}

\item{predicted_y}{The predicted values for y.}

\item{prefix}{The prefix for the dataset in the training or testing.}

\item{target_field}{A string indicating the target field of metadata for regression.}

\item{span}{Controls the amount of smoothing for the default loess smoother in the scatterplot.
Smaller numbers produce wigglier lines, larger numbers produce smoother lines.}

\item{outdir}{The output directory.}
}
\description{
Plot a scatterplot of observed and predicted values from a ranger model.
}
\examples{
set.seed(123)
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
y<- 1:60
rf_model<-rf.out.of.bag(x, y)
plot_obs_VS_pred(y, predicted_y=rf_model$predicted, prefix="train", target_field="age")
}
\author{
Shi Huang
}
