% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_plot_util.R
\name{plot_reg_res_list}
\alias{plot_reg_res_list}
\title{plot_reg_res_list}
\usage{
plot_reg_res_list(reg_res_list, outdir = NULL, plot_height = NULL)
}
\arguments{
\item{reg_res_list}{A list object, the output of the function \code{rf_reg.by_datasets}.}

\item{outdir}{The output directory. If NULL, no file is written.}

\item{plot_height}{The plot height.}
}
\description{
Plot a summary of the performance of a list of regression models
output from \code{rf_reg.by_datasets}.
}
\examples{
df <- data.frame(rbind(t(rmultinom(7, 75, c(.21,.6,.12,.38,.099))),
            t(rmultinom(8, 75, c(.001,.6,.42,.58,.299))),
            t(rmultinom(15, 75, c(.011,.6,.22,.28,.289))),
            t(rmultinom(15, 75, c(.091,.6,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.6,.42,.58,.299)))))
metadata<-data.frame(f_s=factor(c(rep("A", 15), rep("B", 15), rep("A", 15), rep("B", 15))),
                     f_c=factor(c(rep("C", 30), rep("D", 30))),
                     f_d=factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15))),
                     age=c(1:30, 2:31)
                     )
reg_res_list<-rf_reg.by_datasets(df, metadata, s_category="f_c", c_category="age")
plot_reg_res_list(reg_res_list)
}
\seealso{
ranger rf_reg.by_datasets
}
\author{
Shi Huang
}
