% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_reg_plot.R
\name{plot_residuals}
\alias{plot_residuals}
\title{plot_residuals}
\usage{
plot_residuals(y, predicted_y, SampleIDs = NULL, prefix = "train",
  target_field = "value", outdir = NULL)
}
\arguments{
\item{y}{The numeric values for labeling data.}

\item{predicted_y}{The predicted values for y.}

\item{SampleIDs}{The sample ids in the data table.}

\item{prefix}{The prefix for the dataset in the training or testing.}

\item{target_field}{A string indicating the target field of the metadata for regression.}

\item{outdir}{The output directory.}
}
\description{
Plot the residuals of observed and predicted values from a ranger model.
}
\examples{
set.seed(123)
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
y<- 1:60
rf_model<-rf.out.of.bag(x, y)
plot_residuals(y, predicted_y=rf_model$predicted, prefix="train", target_field="age")
}
\author{
Shi Huang
}
