% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_reg_plot.R
\name{plot_train_vs_test}
\alias{plot_train_vs_test}
\title{plot_train_vs_test}
\usage{
plot_train_vs_test(
  train_y,
  predicted_train_y,
  test_y,
  predicted_test_y,
  train_SampleIDs = NULL,
  test_SampleIDs = NULL,
  train_prefix = "train",
  test_prefix = "test",
  train_target_field = "value",
  test_target_field = "value",
  outdir = NULL
)
}
\arguments{
\item{train_y}{The numeric labels for training data.}

\item{predicted_train_y}{The predicted values for training data.}

\item{test_y}{The numeric labels for testing data.}

\item{predicted_test_y}{The predicted values for test data.}

\item{train_SampleIDs}{The sample ids in the train data.}

\item{test_SampleIDs}{The sample ids in the test data.}

\item{train_prefix}{The prefix for the dataset in the training data.}

\item{test_prefix}{The prefix for the dataset in the testing data.}

\item{train_target_field}{A string indicating the target field of the training metadata for regression.}

\item{test_target_field}{A string indicating the target field of the testing metadata for regression.}

\item{outdir}{The output directory.}
}
\description{
Plot the observed and predicted values of both training and testing data from a ranger model.
}
\examples{
set.seed(123)
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
y<- 1:60
newx <- data.frame(rbind(t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
newy <- 31:60
rf_model<-rf.out.of.bag(x, y)
predicted_newy <- predict(rf_model$rf.model, newx)$predictions
plot_train_vs_test(train_y=y, predicted_train_y=rf_model$predicted,
                   test_y=newy, predicted_test_y=predicted_newy,
                   train_target_field="age", test_target_field="age")
}
\author{
Shi Huang
}
