% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{rf_reg.rfe}
\alias{rf_reg.rfe}
\title{rf_reg.rfe}
\usage{
rf_reg.rfe(
  x,
  y,
  nfolds = 5,
  n_features = NULL,
  metric = "MAE",
  tolerance = 0.01,
  recursive = TRUE,
  ntree = 500,
  verbose = FALSE,
  rf_model = NULL,
  ...
)
}
\arguments{
\item{x}{The data frame or data matrix for model training.}

\item{y}{The numeric values for labeling data.}

\item{nfolds}{The number of folds in the cross-validation, or a factor defining customized folds.}

\item{n_features}{A numeric vector of feature set sizes. By default powers of two up to 1024.}

\item{metric}{The regression performance metric used for selection: one of "MAE", "RMSE", "MSE", "MAPE", "Spearman_rho", "R_squared".}

\item{tolerance}{The relative tolerance for choosing the smallest feature set close to the best performance.}

\item{recursive}{A boolean value indicating if the feature ranking is re-computed at each step.}

\item{ntree}{The number of trees.}

\item{verbose}{A boolean value indicating if showing computation status.}

\item{rf_model}{An optional pre-computed \code{rf.cross.validation} or \code{rf.out.of.bag} model on all features.}

\item{...}{Other parameters applicable to \code{ranger}.}
}
\value{
An object of class \code{rf_reg.rfe}.
}
\description{
Recursive feature elimination (RFE) for random forest regression.
See \code{rf_clf.rfe} for details of the procedure.
}
\examples{
set.seed(123)
n_features <- 100
prob_vec <- gtools::rdirichlet(5, sample(n_features))
x <- data.frame(rbind(t(rmultinom(7, 7*n_features, prob_vec[1, ])),
            t(rmultinom(8, 8*n_features, prob_vec[2, ])),
            t(rmultinom(15, 15*n_features, prob_vec[3, ])),
            t(rmultinom(15, 15*n_features, prob_vec[4, ])),
            t(rmultinom(15, 15*n_features, prob_vec[5, ]))))
y<- 1:60
rfe <- rf_reg.rfe(x, y, nfolds=5, metric="MAE", ntree=200)
rfe$top_n_perf
}
\seealso{
plot_reg_feature_selection
}
