% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{save.rf_clf.results.confusion.matrix}
\alias{save.rf_clf.results.confusion.matrix}
\title{save.rf_clf.results.confusion.matrix}
\usage{
save.rf_clf.results.confusion.matrix(result,
  filename = "train.confusion_matrix.xls", outdir = ".")
}
\description{
Print "confusion matrix" file
}
