% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{save.rf_clf.results.importances}
\alias{save.rf_clf.results.importances}
\title{save.rf_clf.results.importances}
\usage{
save.rf_clf.results.importances(result, feature.ids,
  filename = "train.feature_importance_scores.xls", outdir = ".")
}
\description{
Print "feature_importance_scores" file
}
