% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{save.rf_clf.results.mislabeling}
\alias{save.rf_clf.results.mislabeling}
\title{save.rf_clf.results.mislabeling}
\usage{
save.rf_clf.results.mislabeling(result,
  filename = "train.mislabeling.xls", outdir = ".")
}
\description{
Print "mislabeling" file
}
