% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{save.rf_clf.results.probabilities}
\alias{save.rf_clf.results.probabilities}
\title{save.rf_clf.results.probabilities}
\usage{
save.rf_clf.results.probabilities(result,
  filename = "train.cv_probabilities.xls", outdir = ".")
}
\description{
Print "probabilities" file
}
