% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{save.rf_clf.results.summary}
\alias{save.rf_clf.results.summary}
\title{save.rf_clf.results.summary}
\usage{
save.rf_clf.results.summary(result, rf.opts,
  filename = "train.summary.xls", outdir = ".")
}
\description{
Print "summary" file
}
