% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetweenGroup.test.R
\name{BetweenGroup.test}
\alias{BetweenGroup.test}
\title{BetweenGroup.test}
\usage{
BetweenGroup.test(
  x,
  y,
  clr_transform = FALSE,
  p.adj.method = "bonferroni",
  positive_class = NA,
  q_cutoff = 0.2,
  paired = FALSE
)
}
\arguments{
\item{x}{A data.martrix or data.frame including multiple numeric vectors.}

\item{y}{A factor with two or more levels.}

\item{clr_transform}{A logical value indicates if clr transformation before statistical analysis of compositional microbiome data.}

\item{p.adj.method}{A string indicating the p-value correction method.}

\item{positive_class}{A string indicating the specified class in the factor y.}

\item{q_cutoff}{A number indicating the cutoff of q values after fdr correction.}

\item{paired}{A logical indicating if paired between-group comparison is desired.
For paired tests, the samples of the two groups should be in the same order of pairs.}
}
\value{
A data.frame including descriptive statistics of all samples and each group,
the log2 fold changes (\code{mean_logfc}, \code{median_logfc}, \code{generalized_logfc}),
p values and adjusted p values of parametric and non-parametric tests, and the enrichment (\code{Enr}) of each feature.
}
\description{
It runs standard univarate statistical tests (such as t.test, wilcox.test,
oneway.test, kruskal.test)  for all variates (columns) in the data.frame/data.matrix.
}
\examples{
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
# clr_x<-compositions::clr(x)
y<-factor(c(rep("A", 30), rep("B", 30)))
y1<-factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15)))
system.time(BetweenGroup.test(x, y, clr_transform=FALSE))
system.time(BetweenGroup.test(x, y, clr_transform=TRUE))
system.time(BetweenGroup.test(x, y1, clr_transform=TRUE))
x_ <- data.frame(rbind(t(rmultinom(7, 7500, rep(c(.201,.5,.02,.18,.099), 100))),
            t(rmultinom(8, 7500, rep(c(.201,.4,.12,.18,.099), 100))),
            t(rmultinom(15, 7500, rep(c(.011,.3,.22,.18,.289), 100))),
            t(rmultinom(15, 7500, rep(c(.091,.2,.32,.18,.209), 100))),
            t(rmultinom(15, 7500, rep(c(.001,.1,.42,.18,.299), 100)))))
y_<-factor(c(rep("A", 30), rep("B", 30)))
y_1<-factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15)))
system.time(BetweenGroup.test(x_, y_))
system.time(BetweenGroup.test(x_, y_1))
}
\seealso{
clr, t.test, wilcox.test, oneway.test, kruskal.test
}
\author{
Shi Huang
}
