% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_reg_plot.R
\name{calc_rel_predicted}
\alias{calc_rel_predicted}
\title{calc_rel_predicted}
\usage{
calc_rel_predicted(train_y, predicted_train_y, train_SampleIDs = NULL,
  test_y = NULL, predicted_test_y = NULL, test_SampleIDs = NULL,
  train_prefix = "train", test_prefix = "test",
  train_target_field = "y", test_target_field = NULL, outdir = NULL)
}
\arguments{
\item{train_y}{The numeric labels for training data.}

\item{predicted_train_y}{The predicted values for training data.}

\item{train_SampleIDs}{The sample ids in the train data.}

\item{test_y}{The numeric labels for testing data.}

\item{predicted_test_y}{The predicted values for test data.}

\item{test_SampleIDs}{The sample ids in the test data.}

\item{train_prefix}{The prefix for the dataset in the training data.}

\item{test_prefix}{The prefix for the dataset in the testing data.}

\item{train_target_field}{A string indicating the target field of the training metadata for regression.}

\item{test_target_field}{A string indicating the target field of the testing metadata for regression.}

\item{outdir}{The output directory.}
}
\description{
Calculate the relative predicted values to the spline fit.
}
\examples{
set.seed(123)
train_x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
train_y<- 1:60
test_x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209)))))
test_y<- 1:45
train_rf_model<-rf.out.of.bag(train_x, train_y)
predicted_test_y<-predict(train_rf_model$rf.model, test_x)$predictions
calc_rel_predicted(train_y, predicted_train_y=train_rf_model$predicted)
calc_rel_predicted(train_y, predicted_train_y=train_rf_model$predicted, train_SampleIDs=NULL,
                   test_y, predicted_test_y, test_SampleIDs=NULL,
                   train_target_field="y",  test_target_field="test_y", outdir=NULL)
}
\author{
Shi Huang
}
