% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BetweenGroup.test.R
\name{Enr_by_q_cutoff}
\alias{Enr_by_q_cutoff}
\title{Enr_by_q_cutoff}
\usage{
Enr_by_q_cutoff(
  test_output,
  desc_stats_by_group_df,
  positive_class = NA,
  q_cutoff = 0.05
)
}
\arguments{
\item{test_output}{The output from function \code{mttest}.}

\item{desc_stats_by_group_df}{The output from function \code{desc_stats_by_group}.}

\item{positive_class}{A string indicating the specified class in the factor y.}

\item{q_cutoff}{A number indicating the cutoff of q values after fdr correction.}
}
\description{
The significance of all features by q values.
}
\examples{
y <-factor(c(rep("A", 30), rep("B", 30)))
y1<-factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15)))
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
test_output<-mttest(x, y, p.adj.method="bonferroni", paired=FALSE)
test_output<-mttest(x, y1, p.adj.method="bonferroni", paired=FALSE)
desc_stats_by_group_df<-desc_stats_by_group(x, y, clr_transform=FALSE)
desc_stats_by_group_df<-desc_stats_by_group(x, y1, clr_transform=FALSE)
Enr_by_q_cutoff(test_output, desc_stats_by_group_df, q_cutoff=0.05, "A")
}
