% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{get.auroc}
\alias{get.auroc}
\title{get.auroc}
\usage{
get.auroc(prob, obs, positive_class)
}
\arguments{
\item{prob}{A two-column numeric matrix of the same # of rows as the length of observed values,
containing the predicted value of each observation.}

\item{obs}{A binary factor vector indicates observed values.}
}
\value{
\code{auroc}
}
\description{
calculates the area under the ROC curve
}
\examples{
obs<-factor(c(rep("A", 31), rep("B", 29)))
pred <- c(runif(30, 0.5, 0.9), runif(30, 0, 0.6))
prob <-data.frame(A=pred, B=1-pred)
get.auroc(obs, prob, positive_class="A")
}
\author{
Shi Huang
}
