% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{get.oob.votes.from.forest}
\alias{get.oob.votes.from.forest}
\title{get.oob.votes.from.forest}
\usage{
get.oob.votes.from.forest(model, x)
}
\arguments{
\item{model}{A object of class \code{rf.out.of.bag} with elements.}

\item{x}{The training data.}
}
\value{
votes
}
\description{
get votes for each class using only out-of-bag predictions from RF
}
\author{
Shi Huang
}
