% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{get.reg.performance}
\alias{get.reg.performance}
\title{get.reg.performance}
\usage{
get.reg.performance(pred_y, y, n_features = NA)
}
\arguments{
\item{pred_y}{The predicted y in the numeric format.}

\item{y}{A response vector for regression.}

\item{n_features}{The number of features.}
}
\value{
A list of regression performance metrics.
}
\description{
Get regression performance
}
\author{
Shi Huang
}
