% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_plot_util.R
\name{id_non_spcf_markers}
\alias{id_non_spcf_markers}
\title{id_non_spcf_markers}
\usage{
id_non_spcf_markers(feature_res, positive_class = "disease",
  other_class = "health", p.adj.method = "BH", outdir = "./")
}
\arguments{
\item{feature_res}{the inheritant output from the function of plot.res_list, rf_clf.comp.summ or rf_clf.by_dataset.summ}

\item{positive_class}{A string indicates one class in the 'c_category' column of metadata.}

\item{p.adj.method}{The p-value correction method, default is "bonferroni".}

\item{outdir}{The outputh directory, default is "./".}

\item{level1}{A string indicates the class of the factor, default is 'disease'}

\item{level2}{A string indicates the class of the factor, default is 'health'}

\item{q_cutoff}{The cutoff of q values for features, the default value is 0.05.}
}
\value{
...
}
\description{
Non-specific features across datasets (at least present in two of datasets): Last update: 20190130
}
\examples{
df <- data.frame(rbind(t(rmultinom(15, 75, c(.21,.6,.12,.38,.099))),
            t(rmultinom(15, 75, c(.011,.6,.22,.28,.289))),
            t(rmultinom(15, 75, c(.091,.6,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.6,.42,.58,.299)))))
df0 <- data.frame(t(rmultinom(60, 300,c(.001,.6,.2,.3,.299)))) # No feature significantly changed in all sub-datasets!
f=factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15)))
comp_group="A"
res <-rf_clf.comp.summ(df, f, comp_group, verbose=FALSE, ntree=500, p_cutoff=0.05, p.adj.method = "bonferroni", q_cutoff=0.05, outdir="./")
feature_res<-res$feature_res
id_non_spcf_markers(feature_res, positive_class="disease", other_class="health", p.adj.method= "BH", outdir="./")
}
\seealso{
ranger
}
\author{
Shi Huang
}
