% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_clf_plot.R
\name{plot_clf_PRC}
\alias{plot_clf_PRC}
\title{plot_PRC}
\usage{
plot_clf_PRC(
  y,
  rf_clf_model,
  positive_class = NA,
  prefix = "train",
  outdir = NULL
)
}
\arguments{
\item{y}{A factor of classes to be used as the true results}

\item{rf_clf_model}{A list object of a random forest model.}

\item{positive_class}{an optional character string for the factor level that corresponds to a "positive" result (if that makes sense for your data).
If there are only two factor levels, the first level will be used as the "positive" result.}

\item{prefix}{The prefix of data set.}

\item{outdir}{The output directory.}
}
\description{
plot_PRC
}
\examples{
set.seed(123)
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
x0 <- data.frame(rbind(t(rmultinom(7, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(8, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289)))))
y<-factor(c(rep("A", 30), rep("B", 30)))
y0<-factor(c(rep("A", 5), rep("B", 55)))
rf_clf_model<-rf.out.of.bag(x, y)
plot_clf_PRC(y, rf_clf_model, positive_class="A", outdir='./A')
plot_clf_PRC(y, rf_clf_model, positive_class="B", outdir='./B')
rf_clf_model0<-rf.out.of.bag(x0, y0)
plot_clf_PRC(y0, rf_clf_model0, positive_class="A", outdir='./A')
plot_clf_PRC(y0, rf_clf_model0, positive_class="B", outdir='./B')
}
\author{
Shi Huang
}
