% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_reg_plot.R
\name{plot_reg_feature_selection}
\alias{plot_reg_feature_selection}
\title{plot_reg_feature_selection}
\usage{
plot_reg_feature_selection(
  x,
  y,
  rf_reg_model,
  metric = "MAE",
  unit = "unit",
  outdir = NULL
)
}
\arguments{
\item{x}{The data frame or data matrix for model training.}

\item{y}{The numeric values for labeling data.}

\item{rf_reg_model}{The rf regression model from \code{rf.out.of.bag}}

\item{metric}{The regression performance metric applied.
This must be one of "MAE", "RMSE", "MSE", "MAPE".}

\item{unit}{The unit of numeric metadata variables that can be printed in the output figure.}

\item{outdir}{The output directory.}
}
\description{
Plot the regression performance against the reduced number of features used in the modeling.
}
\examples{
set.seed(123)
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
y<- 1:60
rf_reg_model<-rf.out.of.bag(x, y)
plot_reg_feature_selection(x, y, rf_reg_model, metric="MAE", outdir=NULL)
}
\author{
Shi Huang
}
