% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_reg_plot.R
\name{plot_test_perf_VS_rand}
\alias{plot_test_perf_VS_rand}
\title{plot_test_perf_VS_rand}
\usage{
plot_test_perf_VS_rand(
  rf_model,
  newx,
  newy,
  prefix = "test",
  target_field = "",
  metric = "MAE",
  permutation = 1000,
  n_features = NA,
  outdir = NULL
)
}
\arguments{
\item{rf_model}{A trained rf model object, which should be generated from /code{rf.cross.validation} or /code{rf.out.of.bag}.}

\item{newx}{A data.matrix or data.frame with the new data for rf model testing.}

\item{newy}{The data label of new data.}

\item{prefix}{The prefix for the dataset in the training or testing.}

\item{target_field}{A string indicating the target field of the metadata for machine-learning analysis.}

\item{metric}{The regression performance metric applied, including MAE, RMSE, MSE, R_squared, Adj_R_squared, or Separman_rho.}

\item{permutation}{The permutation times for a random guess of regression performance.}

\item{n_features}{The number of features in the training data.}

\item{outdir}{The output directory.}
}
\description{
This outputs a histogram and a p-value showing if the performance of a real regression model
significantly better than null models.
}
\examples{
set.seed(123)
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
y<- 1:60

newx <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
newy<- 4:33
rf_model<-rf.out.of.bag(x, y)
p<-plot_perf_VS_rand(x=x, y=y, predicted_y=rf_model$predicted, prefix="train", nfolds=5,
permutation=100, metric="MAE", target_field="age", n_features=5)
p
p_test<-plot_test_perf_VS_rand(rf_model, newx, newy, permutation=100,
                               metric="MAE", target_field="age", n_features=5)
p_test
}
\author{
Shi Huang
}
