% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_RF_util.R
\name{rf.cross.validation}
\alias{rf.cross.validation}
\title{rf.cross.validation}
\usage{
rf.cross.validation(x, y, nfolds = 3, ntree = 500, verbose = FALSE,
  sparse = FALSE, imp_pvalues = FALSE, ...)
}
\arguments{
\item{x}{Training data: data.matrix or data.frame.}

\item{y}{Data label for each row.}

\item{nfolds}{The number of folds in the cross validation. If nfolds > length(y) or nfolds==-1, uses leave-one-out cross-validation.}

\item{ntree}{The number of trees.}

\item{verbose}{Show computation status and estimated runtime.}

\item{imp_pvalues}{If compute both importance score and pvalue for each feature.}
}
\value{
Object of class \code{rf.cross.validation} with elements including a \code{ranger} object and mutiple metrics for model evaluations.
}
\description{
It runs standard random forests with n-folds cross-validation error estimation for both classification and regression using rf.out.of.bag.
}
\examples{

x0 <- data.frame(t(rmultinom(16,160,c(.001,.5,.3,.3,.299))) + 0.65)
x <- data.frame(rbind(t(rmultinom(7, 75, c(.201,.5,.02,.18,.099))),
            t(rmultinom(8, 75, c(.201,.4,.12,.18,.099))),
            t(rmultinom(15, 75, c(.011,.3,.22,.18,.289))),
            t(rmultinom(15, 75, c(.091,.2,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.1,.42,.18,.299)))))
y<-factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15)))
y0<-factor(c(rep("A", 10), rep("B", 30), rep("C", 5), rep("D", 15)))
system.time(rf.cross.validation(x, y, imp_pvalues=FALSE))
system.time(rf.cross.validation(x, y, imp_pvalues=TRUE))
rf.cross.validation(x, y0, imp_pvalues=FALSE)
y<- 1:60
rf.cross.validation(x, y, nfolds=5, imp_pvalues=FALSE)
rf.cross.validation(x, y, nfolds=5, imp_pvalues=TRUE)
}
\seealso{
ranger
}
\author{
Shi Huang
}
