% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_util.R
\name{rf_clf.by_datasets}
\alias{rf_clf.by_datasets}
\title{rf_clf.by_datasets}
\usage{
rf_clf.by_datasets(df, metadata, s_category, c_category,
  positive_class = NA, rf_imp_pvalues = FALSE, clr_transform = TRUE,
  nfolds = 3, verbose = FALSE, ntree = 500, p.adj.method = "BH",
  q_cutoff = 0.05)
}
\arguments{
\item{df}{Training data: a data.frame.}

\item{metadata}{Sample metadata with at least two columns.}

\item{s_category}{A string indicates the category in the sample metadata: a ‘factor’ defines the sample grouping for data spliting.}

\item{c_category}{A indicates the category in the sample metadata as a responsive vector: if a 'factor', rf classification is performed in each of splited datasets.}

\item{positive_class}{A string indicates one class in the 'c_category' column of metadata.}

\item{rf_imp_pvalues}{A boolean value indicating if compute both importance score and pvalue for each feature.}

\item{clr_transform}{A boolean value indicating if the clr-transformation applied.}

\item{nfolds}{The number of folds in the cross validation.}

\item{verbose}{A boolean value indicating if show computation status and estimated runtime.}

\item{ntree}{The number of trees.}

\item{p.adj.method}{The p-value correction method, default is "bonferroni".}

\item{q_cutoff}{The cutoff of q values for features, the default value is 0.05.}
}
\value{
...
}
\description{
It runs standard random forests with oob estimation for classification of
c_category in each the sub-datasets splited by the s_category,
and apply the model to all the other datasets. The output includes
accuracy, auc and Kappa statistics.
}
\examples{
df <- data.frame(rbind(t(rmultinom(7, 75, c(.21,.6,.12,.38,.099))),
            t(rmultinom(8, 75, c(.001,.6,.42,.58,.299))),
            t(rmultinom(15, 75, c(.011,.6,.22,.28,.289))),
            t(rmultinom(15, 75, c(.091,.6,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.6,.42,.58,.299)))))
df0 <- data.frame(t(rmultinom(60, 300,c(.001,.6,.2,.3,.299))))
metadata<-data.frame(f_s=factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15))),
                     f_c=factor(c(rep("C", 7), rep("H", 8), rep("C", 7), rep("H", 8),
                                  rep("C", 7), rep("H", 8), rep("C", 7), rep("H", 8))),
                     f_d=factor(rep(c(rep("a", 5), rep("b", 5), rep("c", 5)), 4)))
rf_clf.by_datasets(df, metadata, s_category='f_s', c_category='f_c', positive_class="C")
rf_clf.by_datasets(df, metadata, s_category='f_s', c_category='f_c',
                   positive_class="C", rf_imp_pvalues=TRUE)
rf_clf.by_datasets(df, metadata, s_category='f_s', c_category='f_d')
rf_clf.by_datasets(df, metadata, s_category='f_s', c_category='f_d', rf_imp_pvalues=TRUE)
}
\seealso{
ranger
}
\author{
Shi Huang
}
