% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_plot_util.R
\name{rf_clf.by_datasets.summ}
\alias{rf_clf.by_datasets.summ}
\title{rf_clf.by_datasets.summ}
\usage{
rf_clf.by_datasets.summ(df, metadata, s_category, c_category,
  positive_class = NA, rf_imp_pvalues = FALSE, nfolds = 3,
  verbose = FALSE, ntree = 500, p_cutoff = 0.05,
  p.adj.method = "bonferroni", q_cutoff = 0.05, outdir = NULL)
}
\arguments{
\item{df}{Training data: a data.frame.}

\item{metadata}{A metadata with at least two categorical variables.}

\item{s_category}{A string indicates the category in the sample metadata: a ‘factor’ defines the sample grouping for data spliting.}

\item{c_category}{A indicates the category in the sample metadata: a 'factor' used as sample label for rf classification in each of splited datasets.}

\item{positive_class}{A string indicates one class in the 'c_category' column of metadata.}

\item{rf_imp_pvalues}{If compute both importance score and pvalue for each feature.}

\item{nfolds}{The number of folds in the cross validation.}

\item{verbose}{Show computation status and estimated runtime.}

\item{ntree}{The number of trees.}

\item{p_cutoff}{The cutoff of p values for features, the default value is 0.05.}

\item{p.adj.method}{The p-value correction method, default is "bonferroni".}

\item{q_cutoff}{The cutoff of q values for features, the default value is 0.05.}

\item{outdir}{The output directory, the default is "./".}
}
\value{
A list includes a summary of rf models in the sub datasets,
all important statistics for each of features, and plots.
}
\description{
A integrated pipeline for \code{rf_clf.by_datasets}. It runs standard random forests with oob estimation for classification of
c_category in each the sub-datasets splited by the s_category. The output includes a summary of rf models in the sub datasets
and all important statistics for each of features.
}
\seealso{
ranger rf_clf.by_datasets
}
