% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_util.R
\name{rf_clf.comps}
\alias{rf_clf.comps}
\title{rf_clf.comps}
\usage{
rf_clf.comps(
  df,
  f,
  comp_group,
  verbose = FALSE,
  clr_transform = TRUE,
  rf_imp_values = FALSE,
  ntree = 500,
  p.adj.method = "bonferroni",
  q_cutoff = 0.05
)
}
\arguments{
\item{df}{Training data: a data.frame.}

\item{f}{A factor in the metadata with at least two levels (groups).}

\item{comp_group}{A string indicates the group in the f}

\item{verbose}{A boolean value indicating if show computation status and estimated runtime.}

\item{clr_transform}{A boolean value indicating if the clr-transformation applied.}

\item{rf_imp_values}{A boolean value indicating if compute both importance score and pvalue for each feature.}

\item{ntree}{The number of trees.}

\item{p.adj.method}{The p-value correction method, default is "bonferroni".}

\item{q_cutoff}{The cutoff of q values for features, the default value is 0.05.}
}
\value{
...
}
\description{
It runs standard random forests with oob estimation for classification of
one level VS all other levels of one category in the datasets.
The output includes a list of rf models for the sub datasets
and all important statistics for each of features.
}
\examples{
df0 <- data.frame(t(rmultinom(60, 300,c(.001,.6,.2,.3,.299))))
df <- data.frame(rbind(t(rmultinom(7, 75, c(.21,.6,.12,.38,.099))),
            t(rmultinom(8, 75, c(.001,.6,.42,.58,.299))),
            t(rmultinom(15, 75, c(.011,.6,.22,.28,.289))),
            t(rmultinom(15, 75, c(.091,.6,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.6,.42,.58,.299)))))
f=factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15)))
comp_group="A"
comps_res<-rf_clf.comps(df, f, comp_group, verbose=FALSE, ntree=500,
                        p.adj.method = "bonferroni", q_cutoff=0.05)
comps_res
}
\seealso{
ranger
}
\author{
Shi Huang
}
