% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_plot_util.R
\name{rf_clf.comps.summ}
\alias{rf_clf.comps.summ}
\title{rf_clf.comps.summ}
\usage{
rf_clf.comps.summ(df, f, comp_group, clr_transform = TRUE, nfolds = 3,
  verbose = FALSE, ntree = 5000, p_cutoff = 0.05,
  p.adj.method = "bonferroni", q_cutoff = 0.05, outdir = "./")
}
\arguments{
\item{df}{Training data: a data.frame.}

\item{f}{A factor in the metadata with at least two levels (groups).}

\item{comp_group}{A string indicates the group in the f}

\item{nfolds}{The number of folds in the cross validation.}

\item{ntree}{The number of trees.}

\item{p.adj.method}{The p-value correction method, default is "bonferroni".}

\item{q_cutoff}{The cutoff of q values for features, the default value is 0.05.}

\item{outdir}{The outputh directory, default is "./".}

\item{positive_class}{A string indicates one class in the 'c_category' column of metadata.}
}
\value{
A list includes a summary of rf models in the sub datasets,
all important statistics for each of features, and plots.
}
\description{
Runs standard random forests with oob estimation for classification of
one level VS all other levels of one category in the datasets.
The output includes a summary of rf models in the sub datasets,
all important statistics for each of features, and plots.
}
\examples{
df <- data.frame(t(rmultinom(60, 300,c(.001,.6,.2,.3,.299))))
f=factor(c(rep("A", 15), rep("B", 15), rep("C", 15), rep("D", 15)))
comp_group="A"
rf_clf.comps.summ(df, f, comp_group, verbose=FALSE, ntree=500, p_cutoff=0.05, p.adj.method = "bonferroni", q_cutoff=0.05, outdir="./")

}
\seealso{
ranger
}
