% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_util.R
\name{rf_clf.pairwise}
\alias{rf_clf.pairwise}
\title{rf_clf.pairwise}
\usage{
rf_clf.pairwise(df, f, nfolds = 3, ntree = 5000, verbose = FALSE)
}
\arguments{
\item{df}{A data matrix or data.frame}

\item{f}{A factor with more than two levels.}

\item{nfolds}{The number of folds in the cross validation.}

\item{ntree}{The number of trees.}

\item{verbose}{The boolean value indicating if the computation status and estimated runtime shown.}
}
\value{
A summary table containing the performance of Random Forest classification models
}
\description{
Perform pairwise rf classfication for a data matrix between all pairs of group levels
}
\examples{
df <- t(rmultinom(16,160,c(.001,.6,.2,.3,.299))) + 0.65
f<-factor(c(rep("A", 4), rep("B", 4), rep("C", 4), rep("D", 4)))
f0<-factor(c(rep("A", 8), rep("B", 8)))
rf_clf.pairwise(df, f)
}
\seealso{
ranger
}
\author{
Shi Huang
}
