% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranger_crossRF_util.R
\name{rf_reg.cross_appl}
\alias{rf_reg.cross_appl}
\title{rf_reg.cross_appl}
\usage{
rf_reg.cross_appl(rf_list, x_list, y_list)
}
\arguments{
\item{rf_list}{A list of rf.model objects from \code{rf.out.of.bag}.}

\item{x_list}{A list of training datasets usually in the format of data.frame.}

\item{y_list}{A list of responsive vector for regression in the training datasets.}
}
\value{
...
}
\description{
Based on pre-computed rf models regressing \code{c_category} in each the sub-datasets splited by the \code{s_category},
perform cross-datasets application of the rf models. The inputs are precalculated
rf regression models, x_list and y_list.
}
\examples{
df <- data.frame(rbind(t(rmultinom(7, 75, c(.21,.6,.12,.38,.099))),
            t(rmultinom(8, 75, c(.001,.6,.42,.58,.299))),
            t(rmultinom(15, 75, c(.011,.6,.22,.28,.289))),
            t(rmultinom(15, 75, c(.091,.6,.32,.18,.209))),
            t(rmultinom(15, 75, c(.001,.6,.42,.58,.299)))))
df0 <- data.frame(t(rmultinom(60, 300,c(.001,.6,.2,.3,.299))))
metadata<-data.frame(f_s=factor(c(rep("A", 30), rep("B", 30))),
                     f_c=factor(c(rep("C", 15), rep("D", 15), rep("E", 15), rep("F", 15))),
                     age=c(1:30, 2:31)
                     )
reg_res<-rf_reg.by_datasets(df, metadata, s_category='f_c', c_category='age')
rf_reg.cross_appl(reg_res, x_list=reg_res$x_list, y_list=reg_res$y_list)
}
\seealso{
ranger
}
\author{
Shi Huang
}
